+++ packages/go_router/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## NEXT
 
-* Updates minimum supported SDK version to Flutter 3.22/Dart 3.4.
+* Updated the minimum supported SDK version to Flutter 3.22/Dart 3.4.
+* Added new top level `onEnter` callback for controlling incoming route navigation.
 
 ## 14.6.2